WIP: more generic start
